bridges and tunnels working now
Got rid of a few restrictions in the original renderoverlay method by just copying it and removing that part. Works now for bridges and tunnels, but overlay is being projected onto the ground instead of showing at the actual segment height. Might need to swap the overlay effect render method, too